Code refactory. Logic divided over multiple files
