+++ client/consensus/sassafras/src/aux_schema.rs
@@ -16,7 +16,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-//! Schema for Sassafras epoch changes in the auxiliary db.
+//! Schema for auxiliary data persistence.
 
 use scale_codec::{Decode, Encode};
